auto format

+++ test/rsc/patterns/FactoryTest.cpp
@@ -33,66 +33,67 @@ using namespace rsc::patterns;
 
 typedef Factory<string, interface> test_factory;
 
-class FactoryTest : public ::testing::Test {
+class FactoryTest: public ::testing::Test {
 protected:
-  test_factory factory;
+    test_factory factory;
 
-  virtual void
-  SetUp() {
-    this->factory.impls().register_("impl_1",  &impl_1::create);
-    this->factory.impls().register_("impl_2",  &impl_2::create);
-    this->factory.impls().register_("impl_failing_constructor",
-				    &impl_failing_constructor::create);
-  }
+    virtual void SetUp() {
+        this->factory.impls().register_("impl_1", &impl_1::create);
+        this->factory.impls().register_("impl_2", &impl_2::create);
+        this->factory.impls().register_("impl_failing_constructor",
+                &impl_failing_constructor::create);
+    }
 };
 
-TEST_F(FactoryTest, testRegistration) {
-  test_factory factory;
+TEST_F(FactoryTest, testRegistration)
+{
+    test_factory factory;
 
-  // Should start out empty
-  EXPECT_EQ(factory.impls().size(), 0);
+    // Should start out empty
+    EXPECT_EQ(factory.impls().size(), size_t(0));
 
-  // Ordinary registration
-  factory.impls().register_("impl_1",  &impl_1::create);
-  factory.impls().register_("impl_2",  &impl_2::create);
-  factory.impls().register_("impl_failing_constructor",
-			    &impl_failing_constructor::create);
-  EXPECT_EQ(factory.impls().size(), 3);
+    // Ordinary registration
+    factory.impls().register_("impl_1", &impl_1::create);
+    factory.impls().register_("impl_2", &impl_2::create);
+    factory.impls().register_("impl_failing_constructor",
+            &impl_failing_constructor::create);
+    EXPECT_EQ(factory.impls().size(), 3);
 
-  // Duplicates
-  factory.impls().register_("duplicate", &impl_1::create);
-  EXPECT_THROW(factory.impls().register_("duplicate", &impl_1::create),
-	       invalid_argument);
+    // Duplicates
+    factory.impls().register_("duplicate", &impl_1::create);
+    EXPECT_THROW(factory.impls().register_("duplicate", &impl_1::create),
+            invalid_argument);
 
-  // Ordinary Unregistration
-  factory.impls().unregister("impl_1");
-  EXPECT_EQ(factory.impls().size(), 3);
+    // Ordinary Unregistration
+    factory.impls().unregister("impl_1");
+    EXPECT_EQ(factory.impls().size(), 3);
 
-  // Invalid Unregistration
-  EXPECT_THROW(factory.impls().unregister("impl_1"),
-	       NoSuchImpl);
+    // Invalid Unregistration
+    EXPECT_THROW(factory.impls().unregister("impl_1"),
+            NoSuchImpl);
 }
 
-TEST_F(FactoryTest, testCreation) {
-  // Non-existent impl
-  EXPECT_THROW(factory.create_inst("impl_does_not_exist"),
-	       NoSuchImpl);
+TEST_F(FactoryTest, testCreation)
+{
+    // Non-existent impl
+    EXPECT_THROW(factory.create_inst("impl_does_not_exist"),
+            NoSuchImpl);
 
-  // Ordinary creation
-  {
-    Properties p;
-    p["string_param"] = string("test");
-    p["float_param"]  = static_cast<float>(1.0);
-    interface* instance = factory.create_inst("impl_1", p);
-    EXPECT_TRUE(instance);
-  }
+    // Ordinary creation
+    {
+        Properties p;
+        p["string_param"] = string("test");
+        p["float_param"] = static_cast<float>(1.0);
+        interface* instance = factory.create_inst("impl_1", p);
+        EXPECT_TRUE(instance);
+    }
 
-  {
-    interface* instance = factory.create_inst("impl_2");
-    EXPECT_TRUE(instance);
-  }
+    {
+        interface* instance = factory.create_inst("impl_2");
+        EXPECT_TRUE(instance);
+    }
 
-  // Constructor error
-  EXPECT_THROW(factory.create_inst("impl_failing_constructor"),
-	       ConstructError);
+    // Constructor error
+    EXPECT_THROW(factory.create_inst("impl_failing_constructor"),
+            ConstructError);
 }
